Добавляет экран при отсутствии предложений.
