+++ src/components/favorites-screen/favorites-screen-connect.js
@@ -1,0 +1,4 @@
+export const mapStateToProps = (state) => ({
+  cities: state.cities,
+  offers: state.offers,
+});
