+++ src/components/main-screen-no-offers/main-screen-no-offers.jsx
@@ -1,0 +1,21 @@
+const MainScreenNoOffers = (props) => {
+  const {currentCity} = props;
+
+  return (
+    <div className="cities__places-container cities__places-container--empty container">
+      <section className="cities__no-places">
+        <div className="cities__status-wrapper tabs__content">
+          <b className="cities__status">No places to stay available</b>
+          <p className="cities__status-description">We could not find any property available at the moment in {currentCity}</p>
+        </div>
+      </section>
+      <div className="cities__right-section"></div>
+    </div>
+  );
+};
+
+MainScreenNoOffers.propTypes = {
+  currentCity: PropTypes.string.isRequired,
+};
+
+export default MainScreenNoOffers;
