+++ src/components/main-screen-with-offers/main-screen-with-offers-connect.js
@@ -1,0 +1,7 @@
+import {filteredCityOffersSelector} from "@store/selectors";
+
+export const mapStateToProps = (state) => ({
+  currentCity: state.currentCity,
+  offersForCity: filteredCityOffersSelector(state),
+  currentSort: state.currentSort,
+});
