+++ src/components/main-screen/main-screen-connect.js
@@ -1,0 +1,6 @@
+import {filteredCityOffersSelector} from "@store/selectors";
+
+export const mapStateToProps = (state) => ({
+  currentCity: state.currentCity,
+  offersForCity: filteredCityOffersSelector(state),
+});
