+++ src/components/places-list/places-list.jsx
@@ -5,60 +5,58 @@ import {offersPropType} from "@utils/prop-types";
 import {PageType} from "@src/const";
 import {getSortedOffers} from "@utils/offers";
 
-class PlacesList extends React.PureComponent {
-  _getComponentByType(type, offer, onActiveCard, outActiveCard) {
-    switch (type) {
-      case PageType.MAIN:
-        return (
-          <MainPlaceCard
-            key={offer.id}
-            offer={offer}
-            onActiveCard={onActiveCard}
-            outActiveCard={outActiveCard}
-          />
-        );
-      case PageType.FAVORITE:
-        return (
-          <FavoritePlaceCard
-            key={offer.id}
-            offer={offer}
-          />
-        );
-      case PageType.NEAR:
-        return (
-          <NearPlaceCard
-            key={offer.id}
-            offer={offer}
-            onActiveCard={onActiveCard}
-            outActiveCard={outActiveCard}
-          />
-        );
-      default:
-        return null;
-    }
+const getComponentByType = (type, offer, onActiveCard, outActiveCard) => {
+  switch (type) {
+    case PageType.MAIN:
+      return (
+        <MainPlaceCard
+          key={offer.id}
+          offer={offer}
+          onActiveCard={onActiveCard}
+          outActiveCard={outActiveCard}
+        />
+      );
+    case PageType.FAVORITE:
+      return (
+        <FavoritePlaceCard
+          key={offer.id}
+          offer={offer}
+        />
+      );
+    case PageType.NEAR:
+      return (
+        <NearPlaceCard
+          key={offer.id}
+          offer={offer}
+          onActiveCard={onActiveCard}
+          outActiveCard={outActiveCard}
+        />
+      );
+    default:
+      return null;
   }
+};
 
-  render() {
-    const {
-      placesListClassName,
-      offersForCity,
-      type,
-      currentSort,
-      onActiveCard,
-      outActiveCard
-    } = this.props;
+const PlacesList = (props) => {
+  const {
+    placesListClassName,
+    offersForCity,
+    type,
+    currentSort,
+    onActiveCard,
+    outActiveCard
+  } = props;
 
-    const sortedOffers = getSortedOffers(currentSort, offersForCity);
+  const sortedOffers = getSortedOffers(currentSort, offersForCity);
 
-    return (
-      <div className={`${placesListClassName}`}>
-        {sortedOffers.map((offer) => (
-          this._getComponentByType(type, offer, onActiveCard, outActiveCard)
-        ))}
-      </div>
-    );
-  }
-}
+  return (
+    <div className={`${placesListClassName}`}>
+      {sortedOffers.map((offer) => (
+        getComponentByType(type, offer, onActiveCard, outActiveCard)
+      ))}
+    </div>
+  );
+};
 
 PlacesList.propTypes = {
   placesListClassName: PropTypes.string.isRequired,
