+++ src/components/room-screen/room-screen-connect.js
@@ -1,0 +1,6 @@
+import {filteredNearOffersSelector} from "@store/selectors";
+
+export const mapStateToProps = (state) => ({
+  nearOffers: filteredNearOffersSelector(state),
+  reviews: state.reviews,
+});
