+++ src/components/sort-form/sort-form-connect.js
@@ -1,0 +1,3 @@
+export const mapStateToProps = (state) => ({
+  currentSort: state.currentSort,
+});
