+++ src/hocs/with-sort-open/with-sort-open.jsx
@@ -46,9 +46,6 @@ const withSortOpen = (Component) => {
     }
   }
 
-  WithSortOpen.propTypes = {
-  };
-
   return WithSortOpen;
 };
